requirements.txt
streamlit>=1.35.0
tensorflow>=2.14.0
pillow>=10.0.0
numpy>=1.24.0
